+++ src/io/github/frostqui/gui/Font.java
@@ -9,7 +9,8 @@ public class Font {
 	private static String charIndex = "ABCDEFGHIJKLM" + //
 			"NOPQRSTUVWXYZ" + //
 			".,-_" + //
-			"0123456789"; //
+			"0123456789" +  //
+			"abcdefghijklmnopqrstuvwxyz";   //
 			
 			
 
@@ -17,7 +18,7 @@ public class Font {
 	}
 
 	public void render(int x, int y, String text, Screen screen) {
-		render(x, y, 0, 0xffffff, text, screen);
+		render(x, y, -3, 0xffffff, text, screen);
 	}
 	
 	public void render(int x, int y, int spacing, int color, String text, Screen screen) {
@@ -27,12 +28,39 @@ public class Font {
 			xOffset += 8 + spacing;
 			int yOffset = 0;
 			char currentChar = text.charAt(i);
-			if (currentChar == 'g' || currentChar == 'y' || currentChar == 'q' || currentChar == 'p' || currentChar == 'j' || currentChar == ',') yOffset = 4;
+		
+			if (currentChar == 'i') {
+				xOffset -= 1;
+			}
+			
+			
+			
+			if (currentChar == 'm') {
+				xOffset += 1;
+			}
+			
+			if (currentChar == 'w') {
+				xOffset += 1;
+			}
+			
+			
+			
+			
+			
+			
+			
+			if (currentChar == 'g' || currentChar == 'y' || currentChar == 'q' || currentChar == 'p' || currentChar == 'j' || currentChar == ',') {
+				yOffset = 2;
+				
+			}
 			if (currentChar == '\n') {
 				xOffset = 0;
 				line++;
 			}
+			
 			int index = charIndex.indexOf(currentChar);
+			
+			
 			
 			if (index == -1) continue;
 			screen.renderTextCharacter(x + xOffset, y + line * 20 + yOffset, characters[index], color, false);