+++ src/io/github/frostqui/gui/Tile.java
@@ -8,12 +8,6 @@ public class Tile {
 	public int x,y;
 
 
-	
-	public static Tile wildgrass1 = new Tile(Sprite.wildgrass1);
-	public static Tile wildgrass2 = new Tile(Sprite.wildgrass2);
-	public static Tile wildgrass3 = new Tile(Sprite.wildgrass3);
-	public static Tile grass = new Tile(Sprite.grass);
-
 	public Tile(Sprite sprite) {
 		this.sprite = sprite;