+++ src/io/github/frostqui/world/Map.java
@@ -5,13 +5,14 @@ import java.util.Random;
 import io.github.frostqui.gui.Screen;
 import io.github.frostqui.gui.Sprite;
 import io.github.frostqui.gui.Tile;
+import io.github.frostqui.world.tiles.GrassTile;
 
 
 
 public class Map {
 
 	public int w, h;
-	public int[] tiles;
+	public GrassTile[] tiles;
 	
 	private Random random;
 	
@@ -23,12 +24,12 @@ public class Map {
 		this.w = w;
 		this.h = h;
 	
-		tiles = new int[w * h];
+		tiles = new GrassTile[w * h];
 		random = new Random();
 		
 		for(int i = 0; i< w / 16; i++) {
 			for(int j = 0; j< h / 16; j++) {
-				tiles[i+j*w] = 120;
+				tiles[i+j*w] = new GrassTile(Sprite.grass);;
 				
 				
 			}
@@ -36,18 +37,6 @@ public class Map {
 			
 		}
 		
-		for(int i = 0; i< w / 16; i++) {
-			for(int j = 0; j< h / 16; j++) {
-				tiles[i+0*w] = 0;					
-				tiles[i+1*w] = 1;	
-				tiles[i+2*w] = 2;	
-				tiles[0+j*w] = 0;					
-					
-				
-			}
-			
-			
-		}
 		
 		
 
@@ -59,9 +48,11 @@ public class Map {
 		
 		for(int i = 0; i< w / 16; i++) {
 			for(int j = 0; j< h / 16; j++) {
-				getTile(i,j).render(i*16,j*16,screen);
+				tiles[i+j*w].render(i*16,j*16,screen);
+				tiles[i+j*w].update();
 			}
 		}
+		
 		
 		
 	
@@ -69,17 +60,7 @@ public class Map {
 
 	}
 	
-	public Tile getTile(int x, int y) {
-		switch(tiles[x + y * w]) {
-		case 0: return Tile.wildgrass1;
-		case 1:return Tile.wildgrass2;
-		case 2: return Tile.wildgrass3; 
-	
-		default: return Tile.grass;
-		}
-		
-	}
-	
+
 	
 
 }